ParameterCollection: shared ResolveAttachedReferences extension, used by Material.New
