+++ sources/engine/Stride.Rendering/Rendering/Material.cs
@@ -74,66 +74,15 @@ namespace Stride.Rendering
 
             var material = result.Material;
 
-            ResolveAttachedReferences(material, content);
-
-            return material;
-        }
-
-        /// <summary>
-        /// Loads the assets that the generated material only holds references to.
-        /// </summary>
-        /// <param name="material">The generated material.</param>
-        /// <param name="content">The content manager to load through, or <c>null</c> to only report.</param>
-        /// <remarks>
-        /// A material feature attaches a reference with <see cref="AttachedReferenceManager.CreateProxyObject{T}(AssetId, string)"/>,
-        /// which makes an empty object and marks it as a proxy. A proxy becomes the real asset in one
-        /// place only, <c>ReferenceSerializer</c>, and only during a content load. The generator does not
-        /// run there, so the material keeps empty objects until they are loaded here.
-        /// </remarks>
-        private static void ResolveAttachedReferences(Material material, ContentManager content)
-        {
+            // A material feature can attach references to content instead of loaded objects (such as the
+            // lookup table of the default specular model), and only a content load resolves those.
+            // The generator runs outside of one, so the references are loaded here.
             foreach (var pass in material.Passes)
             {
-                var parameters = pass.Parameters;
-                var objectValues = parameters.ObjectValues;
-
-                if (objectValues is null)
-                    continue;
-
-                foreach (var keyInfo in parameters.ParameterKeyInfos)
-                {
-                    // Permutation keys share ObjectValues with resource parameters; replacing one here
-                    // would not bump the permutation counter, so only object (resource) keys are handled.
-                    if (keyInfo.Key.Type != ParameterKeyType.Object || !keyInfo.IsResourceParameter || keyInfo.BindingSlot >= objectValues.Length)
-                        continue;
-
-                    var reference = AttachedReferenceManager.GetAttachedReference(objectValues[keyInfo.BindingSlot]);
-                    if (reference is not { IsProxy: true })
-                        continue;
-
-                    if (content is null)
-                    {
-                        Log.Warning($"Material parameter '{keyInfo.Key}' keeps an empty object, because no " +
-                                    $"content manager was given to load '{reference.Url}'. Pass one to " +
-                                    $"{nameof(Material)}.{nameof(New)}, such as Game.Content.");
-                        continue;
-                    }
-
-                    try
-                    {
-                        objectValues[keyInfo.BindingSlot] = content.Load(keyInfo.Key.PropertyType, reference.Url);
-                    }
-                    catch (ContentManagerException exception)
-                    {
-                        // An asset reaches a build only when something in the content references it.
-                        // Nothing references what a material feature attaches here, because that happens
-                        // at run time, so a game that builds all of its materials in code can lack the
-                        // asset. Keep the empty object rather than stop the game, and say what is wrong.
-                        Log.Warning($"Material parameter '{keyInfo.Key}' keeps an empty object, because the " +
-                                    $"asset it references is not in the build: '{reference.Url}'. {exception.Message}");
-                    }
-                }
+                pass.Parameters.ResolveAttachedReferences(content, Log);
             }
+
+            return material;
         }
     }
 }
